GameData Reader Cost Fix. PlayerJump Sound Fix.

+++ Assets/Scripts/Database/GameDataReader.cs
@@ -503,7 +503,7 @@ public class GameDataReader : MonoBehaviour
                 _ammunitionData[i].critChance = myAmmunitionList.Ammunition[i].criticalHitChance;
                 _ammunitionData[i].critDamage = myAmmunitionList.Ammunition[i].criticalHitDamage;
                 _ammunitionData[i].range = myAmmunitionList.Ammunition[i].range;
-                _ammunitionData[i].cost = (int)myAmmunitionList.Ammunition[i].cost;
+                _ammunitionData[i].cost = myAmmunitionList.Ammunition[i].cost;
 				_ammunitionData[i].rarity = myAmmunitionList.Ammunition[i].rarity;
             }
             else
@@ -533,7 +533,7 @@ public class GameDataReader : MonoBehaviour
                 _barrelData[i].critChance = myBarrelList.Barrel[i].criticalHitChance;
                 _barrelData[i].critDamage = myBarrelList.Barrel[i].criticalHitDamage;
                 _barrelData[i].range = myBarrelList.Barrel[i].range;
-                _barrelData[i].cost = (int)myBarrelList.Barrel[i].cost;
+                _barrelData[i].cost = myBarrelList.Barrel[i].cost;
 				_barrelData[i].rarity = myBarrelList.Barrel[i].rarity;
 			}
             else
@@ -563,7 +563,7 @@ public class GameDataReader : MonoBehaviour
                 _gripData[i].critChance = myGripList.Grip[i].criticalHitChance;
                 _gripData[i].critDamage = myGripList.Grip[i].criticalHitDamage;
                 _gripData[i].range = myGripList.Grip[i].range;
-                _gripData[i].cost = (int)myGripList.Grip[i].cost;
+                _gripData[i].cost = myGripList.Grip[i].cost;
 				_gripData[i].rarity = myGripList.Grip[i].rarity;
 			}
             else
@@ -593,7 +593,7 @@ public class GameDataReader : MonoBehaviour
                 _magazineData[i].critChance = myMagazineList.Magazine[i].criticalHitChance;
                 _magazineData[i].critDamage = myMagazineList.Magazine[i].criticalHitDamage;
                 _magazineData[i].range = myMagazineList.Magazine[i].range;
-                _magazineData[i].cost = (int)myMagazineList.Magazine[i].cost;
+                _magazineData[i].cost = myMagazineList.Magazine[i].cost;
 				_magazineData[i].rarity = myMagazineList.Magazine[i].rarity;
                 _magazineData[i].capacity = myMagazineList.Magazine[i].capacity;
             }
@@ -625,7 +625,7 @@ public class GameDataReader : MonoBehaviour
                 _sightData[i].critChance = mySightList.Sight[i].criticalHitChance;
                 _sightData[i].critDamage = mySightList.Sight[i].criticalHitDamage;
                 _sightData[i].range = mySightList.Sight[i].range;
-                _sightData[i].cost = (int)mySightList.Sight[i].cost;
+                _sightData[i].cost = mySightList.Sight[i].cost;
 				_sightData[i].rarity = mySightList.Sight[i].rarity;
             }
             else
@@ -655,7 +655,7 @@ public class GameDataReader : MonoBehaviour
                 _triggerMechanismData[i].critChance = myTriggerMechanismList.TriggerMechanism[i].criticalHitChance;
                 _triggerMechanismData[i].critDamage = myTriggerMechanismList.TriggerMechanism[i].criticalHitDamage;
                 _triggerMechanismData[i].range = myTriggerMechanismList.TriggerMechanism[i].range;
-                _triggerMechanismData[i].cost = (int)myTriggerMechanismList.TriggerMechanism[i].cost;
+                _triggerMechanismData[i].cost = myTriggerMechanismList.TriggerMechanism[i].cost;
 				_triggerMechanismData[i].rarity = myTriggerMechanismList.TriggerMechanism[i].rarity;
 			}
             else
